Assign 0 to UInt256 by default #131
