+++ robonomics_liability/tests/test_messageValidator.py
@@ -2,6 +2,7 @@
 
 import unittest, rostest, sys
 from robonomics_msgs.msg import Result, Offer, Demand, AddedOrderFeedback, AddedPendingTransactionFeedback
+from ethereum_common.msg import UInt256
 from robonomics_msgs import messageValidator
 import testMessages
 
@@ -84,6 +85,62 @@ class TestMessageValidator(unittest.TestCase):
     def test_ValidAddedPendingTransactionFeedback(self):
         f = messageValidator.convertMessage(testMessages.validAddedPendingTransactionFeedbackDict)
         assert isinstance(f, AddedPendingTransactionFeedback)
+
+    ###
+    # test UInt256 messages
+    ###
+    def test_UInt256_comparison(self):
+        u1 = UInt256("100")
+        u2 = UInt256("101")
+        u3 = UInt256("101")
+
+        # test comparisons with UInt256
+        self.assertGreater(u2, u1)
+        self.assertLess(u1, u2)
+
+        self.assertGreaterEqual(u2, u1)
+        self.assertGreaterEqual(u3, u1)
+        self.assertGreaterEqual(u3, u2)
+        self.assertGreaterEqual(u2, u3)
+
+        self.assertLessEqual(u1, u2)
+        self.assertLessEqual(u1, u3)
+        self.assertLessEqual(u3, u2)
+        self.assertLessEqual(u2, u3)
+
+        self.assertEqual(u2, u3)
+
+        # comparison with int
+        self.assertGreater(u2, 100)
+        self.assertLess(u1, 101)
+
+        self.assertGreaterEqual(u2, 100)
+        self.assertGreaterEqual(u3, 100)
+        self.assertGreaterEqual(u3, 101)
+
+        self.assertLessEqual(u1, 101)
+        self.assertLessEqual(u3, 101)
+
+        self.assertEqual(u1, 100)
+
+    def test_UInt256_default_value(self):
+        u = UInt256()
+        print(u)
+        self.assertEqual(0, int(u.uint256))
+
+    def test_UInt256_arithmetic(self):
+        u1 = UInt256("100")
+        u2 = UInt256("101")
+
+        u1 += 1
+        self.assertEqual(u1, u2)
+
+        u1 = UInt256("100")
+        u3 = u1 + 1
+        self.assertEqual(u3, u2)
+
+        u3 = u1 + UInt256("1")
+        self.assertEqual(u3, u2)
 
 
 if __name__ == '__main__':
